Parameter Dictionary
Substantial changes made as the ParameterDictionary object was used in the GUI and model.  Added a function to build a directory name from the current parameters.  This moves code out of the model to where the parameters actually are.
The model*Parameter changes are mainly bug fixes to allow for how the different data types are handled.