More main-file cleanup
Enough for now.  The main thing here was to set up the toDo list in a separate function.  A_Coral_Model still has 636 lines of code and comments, so it's hardly compact, but it's a start.
By the way, the m files in the main directory have a total of 5,596 lines.