Live script for figure generation
This is a start at generating figures from one central script integrating code calls and descriptions.  This is the closest thing I have found in Matlab to a Markdown type of approach.  In use, a button on the right allows plots to be shown inline or on the right.

Within the "Live Editor", Save has an option for export to PDF or HTML.