Remove generated files and relocate non-essential plotting
Many temporary plotting functions had accumulated which were used as diagnostics or to answer side question.  These clutter the main directory, so they have been moved to a new "Plot_checks" folder.  If any turn out to be useful during actual runs, they can be called with the new path or moved back to the main source directory.

The deleted files are generally obsolete, and can be obtained from the 0.12.1 release if needed.