Deleting obsolete files, renaming functions
Some functions and scripts are just throwaways that are not worth documenting.  Also, I'm renaming functions to start with lower case.  Although MATLAB conventions are to use lowercasefunctionnames with no break between words or names_with_underscores, most of our functions are in camelCase, so I'm going to be consistent with that.