More function renaming
This is going to be really tedious, but software people can be sticklers for consistency.  The worst part is that on case-insensitive filesystems Github won't let you rename a file when the only change is letter case.  I'm going to double the first letter and then re-commit with the final intended name in cases where this comes up.