Move climate data into repository
In many ways it made more sense to store the climate data totally separately, especially to avoid duplication.  Nevertheless, it will be easier for new users to get the code and data together, and for me to document it.
While doing this, I chose clearer (but not perfect) variable names for the paths.