Changes to the main model function
Dozens of changes to update the ways symbionts and corals are associated, and to pull symbiont population histories back from the parallel processes so relative population sizes can be analyzed.
This also allows growthPenalty to be passed as a parameter to the iteration functions so it may be changed without recompiling.