Figure generation
This project has accumulated a large number of figure possibilities, some mean for publication and some as diagnostics.  Many of these need to be removed as obsolete, but at least this includes the latest versions as of March 2021.
Note that AllFigures.m generates several of the main figures and tables submitted for publication, but it definitely does not generate ALL the figures.